tests/submission/PhilipTaylor/canvas/2d.text.measure.width.basic.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: 2d.text.measure.width.basic</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
<style>
@font-face {
  font-family: CanvasTest;
  src: url("../fonts/CanvasTest.ttf");
}
</style>
</head><body class="show_output">

<h1><a href="index.2d.html">2d</a>.<a href="index.2d.text.html">text</a>.<a href="index.2d.text.measure.html">measure</a>.<a href="index.2d.text.measure.width.html">width</a>.basic</h1>
<p class="desc"></p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.2d.text.measure">2d.text.measure</a>
</li></ul>
</div>

<span style="font-family: CanvasTest; position: absolute; visibility: hidden">A</span>
<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("");
_addTest(function(canvas, ctx) {

deferTest();
setTimeout(wrapFunction(function () {
    ctx.font = '50px CanvasTest';
    _assertSame(ctx.measureText('A').width, 50, "ctx.measureText('A').width", "50");
    _assertSame(ctx.measureText('AA').width, 100, "ctx.measureText('AA').width", "100");
    _assertSame(ctx.measureText('ABCD').width, 200, "ctx.measureText('ABCD').width", "200");

    ctx.font = '100px CanvasTest';
    _assertSame(ctx.measureText('A').width, 100, "ctx.measureText('A').width", "100");
}), 500);


});
</script>

<div id="log"></div>
</body></html>
